fix everything except getsentry
