+++ tests/sentry/incidents/endpoints/validators/test_validators.py
@@ -486,7 +486,7 @@ class TestMetricAlertsDetectorValidator(BaseValidatorTest):
 
         # Create a not-metric detector
         self.create_detector(
-            project_id=self.project.id,
+            project=self.project,
             name="Error Detector",
             status=ObjectStatus.ACTIVE,
         )
